Move composition resource interfaces and unstructured implementations to crossplane-runtime.
