+++ pkg/controller/apiextensions/composite/engine.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package composed
+package composite
 
 import (
 	"github.com/pkg/errors"
